renaming svg icons files
