+++ src/components/icons/external-link.tsx
